Move test file of problem C
